+++ src/domain/page/pageHeader/PageHeaderContainer.tsx
@@ -3,14 +3,14 @@ import { Link } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { Button, IconShare, IconAngleRight } from 'hds-react';
 
-import { IconPerson } from '../../common/icons';
-import Header from '../../common/header/Header';
-import HelsinkiLogo from '../../common/helsinkiLogo/HelsinkiLogo';
-import Dropdown from '../../common/dropdown/Dropdown';
-import List from '../../common/list/List';
-import ListItem from '../../common/list/ListItem';
-import { useCurrentUser } from '../auth/hooks';
-import authService from '../auth/authService';
+import { IconPerson } from '../../../common/icons';
+import Header from '../../../common/header/Header';
+import HelsinkiLogo from '../../../common/helsinkiLogo/HelsinkiLogo';
+import Dropdown from '../../../common/dropdown/Dropdown';
+import List from '../../../common/list/List';
+import ListItem from '../../../common/list/ListItem';
+import { useCurrentUser } from '../../auth/hooks';
+import authService from '../../auth/authService';
 
 const PageHeaderContainer: React.SFC = () => {
   const { t } = useTranslation();
